src(gps): add gps driver functionality
The following methods have been implemented with the GPS driver and
tested:
* read
* connect_to_satellites
* is_connected_to_satellites
* vehicle_state
* update_cache

+++ src/sensor_data_structures/gps_data.h
@@ -22,9 +22,9 @@ struct gps_data {
         uint8_t day;
         uint8_t hour;
         uint8_t minute;
-        uint8_t second;
-        uint8_t millisecond;
-        float latitude;
-        float longitude;
-        float speed;
+        uint16_t second;
+        uint16_t millisecond;
+        int32_t latitude;
+        int32_t longitude;
+        int32_t speed;
 };
